Moved ProjectPage into its own folder.
